requirements.txt
colorthief
fiona
geopandas
matplotlib
numpy
pandas
Pillow
plotly
quarto-cli